backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Pillow==10.2.0
